Add Keychain usage description and downgrade flutter_secure_storage
Add `NSKeychainUsageDescription` to the iOS Info.plist to provide a description for using Keychain. This is required for apps that use the Keychain services.

Downgrade the `flutter_secure_storage` package from version 9.2.4 to 8.1.0, along with its platform-specific dependencies (`flutter_secure_storage_windows` from 3.1.2 to 2.1.1). This change ensures compatibility with the current project setup.